refactor: moves itemschema into seperate file & renames table components.
